requirements.txt
astroid==2.2.5
beautifulsoup4==4.6.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.5
django-cogwheels==0.2
django-debug-toolbar==2.0
django-modelcluster==4.4
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.10.3
draftjs-exporter==2.1.6
html5lib==1.0.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.1.0
psycopg2==2.8.3
pylint==2.3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.3
wagtail==2.6.1
wagtail-cache==0.5.2
wagtail-condensedinlinepanel==0.5.2
wagtailmenus==2.13.1
webencodings==0.5.1
Willow==1.1
wrapt==1.11.2
